thats a lot of errors

+++ server/db/db.sqlite.ts
@@ -2,15 +2,98 @@
 the implementation of the database. this is the nervous system of our game.
 */
 
-import { DatabaseSync } from "node:sqlite";
+import { DatabaseSync, SQLInputValue, SQLOutputValue, StatementResultingChanges, StatementSync } from "node:sqlite";
 import { DB, User, Session, CatInst, CatDef, TradeLocal } from "./db.types.ts";
 import { Miss } from "shared/utility.ts";
 import { CatDefJson } from "shared/types.ts";
+
+const util_hash = (str: string, limit: number): number => {
+	let hash = 0;
+	for (let i = 0, len = Math.min(str.length, limit); i < len; i++) {
+		hash = ((hash << 5) - hash + str.charCodeAt(i)) | 0;
+	}
+	return hash;
+};
+
+const util_hash_list = (str_list: Iterable<string>, limit: number, sub: number): number => {
+	let acc = 0, left = limit;
+	for (const s of str_list) {
+		const diff = Math.min(left, sub);
+		acc = acc + util_hash(s, diff) | 0;
+		left -= diff;
+		if (left <= 0) {
+			break;
+		}
+	}
+	return acc;
+};
+
+class Cache {
+	constructor(db: DatabaseSync) {
+		this.#db = db;
+		this.#internal = new Map();
+	}
+
+	#db: DatabaseSync;
+	#internal: Map<number, (readonly [number, number, string, StatementSync])[]>;
+
+	cache(str: TemplateStringsArray, input_count: number): StatementSync {
+		const hash = util_hash_list(str, 32, 8);
+
+		let l1 = this.#internal.get(hash);
+		if (l1 === undefined) {
+			l1 = [];
+			this.#internal.set(hash, l1);
+		}
+
+		let collect_cache: string | undefined;
+		const collect = () => {
+			if (collect_cache !== undefined) {
+				return collect_cache;
+			}
+			let acc = "", left = input_count;
+			str.forEach(x => (acc += x, acc += (left-- > 0) ? "(?)" : ""));
+			collect_cache = acc;
+			return collect_cache;
+		};
+		
+		let l2;
+		for (const lx of l1) {
+			if (lx[0] === str.length && lx[1] === input_count && lx[2] === collect()) {
+				l2 = lx;
+				break;
+			}
+		}
+
+		if (l2 === undefined) {
+			const str = collect();
+			const statement = this.#db.prepare(str);
+			l2 = [str.length, input_count, str, statement] as const;
+			l1.push(l2);
+		}
+
+		return l2[3];
+	}
+
+	run(str: TemplateStringsArray, ...input: SQLInputValue[]): StatementResultingChanges {
+		return this.cache(str, input.length).run(...input);
+	}
+
+	all(str: TemplateStringsArray, ...input: SQLInputValue[]): Record<string, SQLOutputValue>[] {
+		return this.cache(str, input.length).all(...input);
+	}
+
+	get(str: TemplateStringsArray, ...input: SQLInputValue[]): Record<string, SQLOutputValue> | undefined {
+		return this.cache(str, input.length).get(...input);
+	}
+}
 
 export class DBSql implements DB {
 	constructor(path: string) {
 		this.db = new DatabaseSync(path + '/db.sql');
 
+		this.sql = new Cache(this.db);
+		
 		this.db.exec(`
 			PRAGMA foreign_keys = ON;
 			
@@ -18,8 +101,15 @@ export class DBSql implements DB {
 				id INTEGER PRIMARY KEY AUTOINCREMENT,
 				username TEXT UNIQUE NOT NULL,
 				password TEXT NOT NULL,
-				settings TEXT NOT NULL,
 				tokens INTEGER NOT NULL
+			);
+
+			CREATE TABLE IF NOT EXISTS settings (
+				user_id INTEGER NOT NULL,
+				key TEXT NOT NULL,
+				value TEXT NOT NULL,
+				PRIMARY KEY (user_id, key),
+				FOREIGN KEY (user_id) REFERENCES users(id) ON DELETE CASCADE
 			);
 
 			CREATE TABLE IF NOT EXISTS sessions (
@@ -58,16 +148,17 @@ export class DBSql implements DB {
 	}
 
 	db: DatabaseSync;
+	sql: Cache;
 
-	async user_new(username: string, password: string, settings: string): Promise<number | Miss<'internal' | 'exists'>> {
+	async user_new(username: string, password: string): Promise<number | Miss<'internal' | 'exists'>> {
 		let result;
+
 		try {
-			result = this.db.prepare(`
+			result = this.sql.run `
 				INSERT INTO users
-					(username, password, settings, tokens)
+					(username, password, tokens)
 				VALUES
-					(?, ?, ?, ?);
-			`).run(username, password, settings, 5);
+					(${username}, ${password}, ${100});`;
 		}
 		catch (_e) {
 			return new Miss('exists', `username '${username}' already exists`);
@@ -80,10 +171,9 @@ export class DBSql implements DB {
 		let result;
 
 		try {
-			result = this.db.prepare(`
+			result = this.sql.get `
 				SELECT * FROM users
-				WHERE username = (?);
-			`).get(username);
+				WHERE username = (${username});`;
 		}
 		catch (e) {
 			console.error(e);
@@ -101,10 +191,9 @@ export class DBSql implements DB {
 		let result;
 
 		try {
-			result = this.db.prepare(`
+			result = this.sql.get `
 				SELECT * FROM users
-				WHERE id = (?);
-			`).get(user_id);
+				WHERE id = (${user_id});`;
 		}
 		catch (e) {
 			console.error(e);
@@ -120,16 +209,15 @@ export class DBSql implements DB {
 
 	async user_set(user: User): Promise<null | Miss<"internal" | "conflict">> {
 		try {
-			this.db.prepare(`
+			this.sql.run `
 				UPDATE users
 				SET
-					username = (?),
-					password = (?),
-					settings = (?),
-					tokens = (?)
+					username = (${user.username}),
+					password = (${user.password}),
+					settings = (${user.settings}),
+					tokens = (${user.tokens})
 				WHERE
-					id = (?);
-			`).run(user.username, user.password, user.settings, user.tokens, user.id);
+					id = (${user.id});`;
 		}
 		catch (_e) {
 			return new Miss('conflict', `username '${user.username}' already exists`);
@@ -140,10 +228,7 @@ export class DBSql implements DB {
 
 	async user_delete(user_id: number): Promise<null | Miss<"internal" | "not_found">> {
 		try {
-			this.db.prepare(`
-				DELETE FROM users
-				WHERE id = (?);
-			`).run(user_id);
+			this.sql.run `DELETE FROM users WHERE id = (${user_id});`;
 		}
 		catch (e) {
 			console.error(e);
@@ -153,15 +238,67 @@ export class DBSql implements DB {
 		return null;
 	}
 
+	async settings_get(user_id: number, key: string, def?: string): Promise<string | undefined | Miss<"internal">> {
+		let result;
+
+		try {
+			result = this.sql.get `SELECT value FROM settings WHERE user_id = ${user_id} AND key = ${key}`;
+		}
+		catch (e) {
+			console.error(e);
+			return new Miss('internal', `unknown internal error`);
+		}
+
+		if (result === undefined && def !== undefined) {
+			const result_set = await this.settings_set(user_id, key, def);
+			if (result_set instanceof Miss) {
+				return result_set;
+			}
+
+			return def;
+		}
+
+		return result !== undefined ? result['value'] as string : undefined;
+	}
+
+	async settings_set(user_id: number, key: string, value: string): Promise<null | Miss<"internal">> {
+		try {
+			this.sql.run `
+				INSERT INTO settings
+					(user_id, key, value)
+				VALUES
+					(${user_id}, ${key}, ${value})
+				ON CONFLICT (user_id, key) DO UPDATE SET
+					value = excluded.value;`;
+		}
+		catch (e) {
+			console.error(e);
+			return new Miss('internal', `unknown internal error`);
+		}
+
+		return null;
+	}
+
+	async settings_list(user_id: number): Promise<[string, string][] | Miss<"internal">> {
+		let result;
+
+		try {
+			result = this.sql.all `SELECT * FROM settings WHERE user_id = ${user_id};`;
+		}
+		catch (e) {
+			console.error(e);
+			return new Miss('internal', `unknown internal error`);
+		}
+
+		return result.values().map(x => [x['key'] as string, x['value'] as string] as [string, string]).toArray();
+	}
+
 	async session_new(user_id: number, expires: number): Promise<string | Miss<'internal'>> {
 		const now = Date.now();
 
 		// purge old sessions
 		try {
-			this.db.prepare(`
-				DELETE FROM sessions
-				WHERE date_expire <= (?);
-			`).run(now);
+			this.sql.run `DELETE FROM sessions WHERE date_expire <= (${now});`;
 		}
 		catch (_e) {
 			return new Miss('internal', `sessions could not be purged`);
@@ -172,12 +309,11 @@ export class DBSql implements DB {
 		const time = now + 1000 * 60 * 60 * 60 * expires;
 
 		try {
-			this.db.prepare(`
+			this.sql.run `
 				INSERT INTO sessions
 					(id, csrf, user_id, date_expire)
 				VALUES
-					(?, ?, ?, ?);
-			`).run(session_id, csrf, user_id, time);
+					(${session_id}, ${csrf}, ${user_id}, ${time});`;
 		}
 		catch (_e) {
 			return new Miss('internal', `session could not be deleted`)
@@ -190,10 +326,9 @@ export class DBSql implements DB {
 		let result;
 
 		try {
-			result = this.db.prepare(`
+			result = this.sql.get `
 				SELECT * FROM sessions
-				WHERE sessions.id = (?) AND sessions.date_expire > (?);
-			`).get(session_id, Date.now());
+				WHERE sessions.id = ${session_id} AND sessions.date_expire > ${Date.now()};`;
 		}
 		catch (e) {
 			console.error(e);
@@ -209,10 +344,7 @@ export class DBSql implements DB {
 
 	async session_delete(session_id: string): Promise<null | Miss<'internal' | 'not_found'>> {
 		try {
-			this.db.prepare(`
-				DELETE FROM sessions
-				WHERE id = (?);
-			`).run(session_id);
+			this.sql.run `DELETE FROM sessions WHERE id = (${session_id});`;
 		}
 		catch (_e) {
 			return new Miss('not_found', `session does not exist`);
@@ -224,15 +356,14 @@ export class DBSql implements DB {
 	async catdefs_sync(catdefs: CatDefJson[]): Promise<null | Miss<"internal">> {
 		for (const catdef of catdefs) {
 			try {
-				this.db.prepare(`
+				this.sql.run `
 					INSERT INTO catdefs
 						(key, name, rarity)
 					VALUES
-						(?, ?, ?)
+						(${catdef.key}, ${catdef.name}, ${catdef.rarity})
 					ON CONFLICT (key) DO UPDATE SET
 						name = excluded.name,
-						rarity = excluded.rarity;
-				`).run(catdef.key, catdef.name, catdef.rarity);
+						rarity = excluded.rarity;`;
 			}
 			catch (_e) {
 				console.error(_e);
@@ -247,9 +378,7 @@ export class DBSql implements DB {
 		let result;
 
 		try {
-			result = this.db.prepare(`
-				SELECT * FROM catdefs;
-			`).all();
+			result = this.sql.all `SELECT * FROM catdefs;`;
 		}
 		catch (_e) {
 			console.error(_e);
@@ -263,10 +392,7 @@ export class DBSql implements DB {
 		let result;
 
 		try {
-			result = this.db.prepare(`
-				SELECT * FROM catdefs
-				WHERE id = (?);
-			`).get(catdef_id);
+			result = this.sql.get `SELECT * FROM catdefs WHERE id = (${catdef_id});`;
 		}
 		catch (_e) {
 			console.error(_e);
@@ -280,12 +406,11 @@ export class DBSql implements DB {
 		let result;
 
 		try {
-			result = this.db.prepare(`
+			result = this.sql.run `
 				INSERT INTO catinsts
 					(catdef_id, original_user_id, owner_user_id, name)
 				VALUES
-					(?, ?, ?, ?);
-			`).run(catdef_id, user_id, user_id, "cat");
+					(${catdef_id}, ${user_id}, ${user_id}, ${"cat"});`;
 		}
 		catch (e) {
 			console.error(e);
@@ -299,10 +424,7 @@ export class DBSql implements DB {
 		let result;
 
 		try {
-			result = this.db.prepare(`
-				SELECT * FROM catinsts
-				WHERE id = (?);
-			`).get(catinst_id);
+			result = this.sql.get `SELECT * FROM catinsts WHERE id = (${catinst_id});`;
 		}
 		catch (e) {
 			console.error(e);
@@ -318,14 +440,13 @@ export class DBSql implements DB {
 
 	async catinst_set(catinst: CatInst): Promise<null | Miss<"internal">> {
 		try {
-			this.db.prepare(`
+			this.sql.run `
 				UPDATE catinsts
 				SET
-					owner_user_id = (?),
-					name = (?)
+					owner_user_id = (${catinst.owner_user_id}),
+					name = (${catinst.name})
 				WHERE
-					id = (?);
-			`).run(catinst.owner_user_id, catinst.name, catinst.id);
+					id = (${catinst.id});`;
 		}
 		catch (e) {
 			console.error(e);
@@ -337,10 +458,7 @@ export class DBSql implements DB {
 
 	async catinst_delete(catinst_id: number): Promise<null | Miss<"internal" | "not_found">> {
 		try {
-			this.db.prepare(`
-				DELETE FROM catinsts
-				WHERE id = (?);
-			`).run(catinst_id);
+			this.sql.run `DELETE FROM catinsts WHERE id = (${catinst_id});`;
 		}
 		catch (e) {
 			console.error(e);
@@ -354,12 +472,11 @@ export class DBSql implements DB {
 		let result;
 
 		try {
-			result = this.db.prepare(`
+			result = this.sql.all `
 				SELECT catinsts.* FROM catinsts
 				JOIN catdefs ON catinsts.catdef_id = catdefs.id
-				WHERE catinsts.owner_user_id = (?) AND (catinsts.name LIKE '%' || (?) || '%' OR catdefs.name LIKE '%' || (?) || '%')
-				LIMIT (?) OFFSET (?);
-			`).all(user_id, query, query, limit, offset);
+				WHERE catinsts.owner_user_id = (${user_id}) AND (catinsts.name LIKE '%' || (${query}) || '%' OR catdefs.name LIKE '%' || (${query}) || '%')
+				LIMIT (${limit}) OFFSET (${offset});`;
 		}
 		catch (e) {
 			console.error(e);
@@ -373,12 +490,11 @@ export class DBSql implements DB {
 		let result;
 
 		try {
-			result = this.db.prepare(`
+			result = this.sql.run `
 				INSERT INTO tradelocal
 					(peer_x_id, peer_y_id, catinst_x_id, catinst_y_id)
 				VALUES
-					(?, ?, ?, ?);
-			`).run(creator_user_id, with_user_id, creator_catinst_id, with_catinst_id);
+					(${creator_user_id}, ${with_user_id}, ${creator_catinst_id}, ${with_catinst_id});`;
 		}
 		catch (e) {
 			console.error(e);
@@ -392,10 +508,7 @@ export class DBSql implements DB {
 		let result;
 
 		try {
-			result = this.db.prepare(`
-				SELECT * FROM tradelocal
-				WHERE id = (?);
-			`).get(tradelocal_id);
+			result = this.sql.get `SELECT * FROM tradelocal WHERE id = (${tradelocal_id});`;
 		}
 		catch (e) {
 			console.error(e);
@@ -413,10 +526,7 @@ export class DBSql implements DB {
 		let result;
 
 		try {
-			result = this.db.prepare(`
-				DELETE FROM tradelocal
-				WHERE id = (?);
-			`).run(tradelocal_id);
+			result = this.sql.run `DELETE FROM tradelocal WHERE id = (${tradelocal_id});`;
 		}
 		catch (e) {
 			console.error(e);
